生成物冲突（tests/agent_eval/eval_cases.py / docs/testing/mibao-verification-cases.md）
按 §8/§11.2 处理：不手改，以合并后的 .github/cases/ 为单一源重渲染
（288 条 = 本分支 PP-009 + main #3671 的 OR-028，两侧都保留）。

Refs #3672

+++ backend/admin-api/src/main/java/com/migao/admin/dto/OrderDetailResponse.java
@@ -134,9 +134,9 @@ public class OrderDetailResponse {
         private String skuCode;
 
         /**
-         * 数量
+         * 数量（口径按计价方式：per_meter=米数 / per_set=1 / per_area=宽×高㎡，可为小数）
          */
-        private Integer quantity;
+        private BigDecimal quantity;
 
         /**
          * 单价
@@ -196,9 +196,9 @@ public class OrderDetailResponse {
         private BigDecimal unitPrice;
 
         /**
-         * 数量
+         * 数量（口径按计价方式：per_meter=米数 / per_set=1 / per_area=宽×高㎡，可为小数）
          */
-        private Integer quantity;
+        private BigDecimal quantity;
 
         /**
          * 金额 = unitPrice * quantity
